+++ packages/terra-core-docs/src/terra-dev-site/doc/section-header/example/LongTitleAccordionSectionHeader.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import SectionHeaderExampleTemplate from './SectionHeaderExampleTemplate';
+import AccoordianExampleTemplate from './AccordionExampleTemplate';
 
 const LongTitleAccordionSectionHeader = () => {
   const [isOpen, setIsOpen] = useState(false);
@@ -9,13 +9,15 @@ const LongTitleAccordionSectionHeader = () => {
   };
 
   const sectionHeaderProps = {
-    title: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteu',
     isOpen,
     onClick: handleClick,
   };
 
   return (
-    <SectionHeaderExampleTemplate title="Long Title Accordion Section Header" exampleProps={sectionHeaderProps} />
+    <AccoordianExampleTemplate title="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteu" heading="Long Title Accordion Section Header" sectionHeaderAttrs={sectionHeaderProps}>
+      <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteu</p>
+      <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteu</p>
+    </AccoordianExampleTemplate>
   );
 };