+++ packages/terra-core-docs/src/terra-dev-site/test/section-header/DefaultSectionHeader.test.jsx
@@ -1,4 +1,4 @@
 import React from 'react';
 import SectionHeader from 'terra-section-header';
 
-export default () => <SectionHeader title="Default" className="defaultSectionHeader" />;
+export default () => <SectionHeader text="Default" className="defaultSectionHeader" />;