+++ packages/terra-alert/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 4.61.0 - (December 7, 2022)
+
+* Changed
+  * Minor dependency version bump
 
 ## 4.60.7 - (July 5, 2022)