+++ packages/terra-base/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 5.49.0 - (December 7, 2022)
+
+* Changed
+  * Minor dependency version bump
 
 ## 5.48.1 - (January 4, 2022)