+++ packages/terra-breakpoints/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 2.37.0 - (December 7, 2022)
+
+* Changed
+  * Minor dependency version bump
 
 ## 2.36.0 - (October 8, 2021)