+++ packages/terra-divider/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 3.37.0 - (December 7, 2022)
+
+* Added
+  * Added a new props `id`, `level` to the divider.
 
 ## 3.36.1 - (December 14, 2021)