+++ packages/terra-dropdown-button/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 1.30.0 - (December 7, 2022)
+
+* Changed
+  * Minor dependency version bump
 
 ## 1.29.1 - (December 14, 2021)