+++ packages/terra-form-field/CHANGELOG.md
@@ -2,8 +2,15 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 4.21.0 - (December 7, 2022)
+
+* Changed
+  * Set the hideRequired prop to be private
+  * Set the isLabelHidden prop to be private
+  * Set an asterisk denoting required field to be aria-hidden
+
+* Removed
+  * Removed aria-live attribute from error message.
 
 ## 4.20.5 - (July 5, 2022)