+++ packages/terra-form-fieldset/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 2.64.0 - (December 7, 2022)
+
+* Changed
+  * Minor dependency version bump
 
 ## 2.63.5 - (July 5, 2022)