+++ packages/terra-form-input/CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 4.18.0 - (December 7, 2022)
+
+* Changed
+  * Set the hideRequired prop to be private
+  * Set the isLabelHidden prop to be private
+  * Set the ariaLabel prop to be private
 
 ## 4.17.5 - (July 5, 2022)