+++ packages/terra-form-radio/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 4.31.0 - (December 7, 2022)
+
+* Changed
+  * Minor dependency version bump
 
 ## 4.30.1 - (February 15, 2022)