+++ packages/terra-form-select/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 6.36.0 - (December 7, 2022)
+
+* Changed
+  * Minor dependency version bump
 
 ## 6.35.5 - (July 5, 2022)