+++ packages/terra-form-textarea/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 5.20.0 - (December 7, 2022)
+
+* Changed
+  * Minor dependency version bump
 
 ## 5.19.5 - (July 5, 2022)