+++ packages/terra-heading/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 4.49.0 - (December 7, 2022)
+
+* Changed
+  * Minor dependency version bump
 
 ## 4.48.1 - (December 14, 2021)