+++ packages/terra-html-table/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 1.15.0 - (December 7, 2022)
+
+* Changed
+  * Added ability to add row headers to the table.
 
 ## 1.14.0 - (July 14, 2022)