+++ packages/terra-i18n/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 4.39.0 - (December 7, 2022)
+
+* Changed
+  * Minor dependency version bump
 
 ## 4.38.1 - (January 4, 2022)