+++ packages/terra-list/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 4.52.0 - (December 7, 2022)
+
+* Changed
+  * Sections and subsections updated to add programmatic context to communicate hierarchal relationships
 
 ## 4.51.5 - (July 5, 2022)