+++ packages/terra-profile-image/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 3.42.0 - (December 7, 2022)
+
+* Changed 
+  * Added `role = presentation` for Decorative profile-image.
 
 ## 3.41.1 - (January 4, 2022)