+++ packages/terra-show-hide/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 2.56.0 - (December 7, 2022)
+
+* Changed
+  * Minor dependency version bump
 
 ## 2.55.3 - (July 5, 2022)