+++ packages/terra-spacer/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 3.60.0 - (December 7, 2022)
+
+* Changed
+  * Minor dependency version bump
 
 ## 3.59.2 - (January 4, 2022)