+++ packages/terra-tag/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 2.53.0 - (December 7, 2022)
+
+* Changed
+  * Minor dependency version bump
 
 ## 2.52.3 - (July 5, 2022)