+++ packages/terra-text/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 4.50.0 - (December 7, 2022)
+
+* Changed
+  * Minor dependency version bump
 
 ## 4.49.0 - (February 2, 2022)