+++ packages/terra-toggle-button/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 3.62.0 - (December 7, 2022)
+
+* Changed
+  * Minor dependency version bump
 
 ## 3.61.4 - (July 5, 2022)