+++ packages/terra-toggle-section-header/CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 2.59.0 - (December 7, 2022)
+
+* Changed
+  * Renamed `terra-section-header` title prop to text
+  * Updated snapshots from changes made to terra-section-header
+  * Updated WDIO Screenshots from changes made to terra-section-header
 
 ## 2.58.4 - (July 14, 2022)