+++ packages/terra-toolbar/CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed broken link in documentation.
+## 1.30.0 - (December 7, 2022)
+
+* Added
+  * Role attribute is set as toolbar
+  * Prop `ariaLabel` - String that labels what content the toolbar controls.
+  * Prop `ariaLabelledBy` - String that labels what content (referencing the element id) the toolbar controls.
 
 ## 1.29.0 - (March 15, 2022)